+++ uui/src/lib/List/ListItem.types.d.ts
@@ -5,17 +5,11 @@ export type ListItemSlotProps = Record<string, boolean | undefined>;
 export interface ListItemProps<T extends EventTarget> extends BoxProps<T> {
   text: string;
   secondaryText?: string;
+  textColor?: string;
   selected?: boolean;
   divider?: boolean;
   disabled?: boolean;
   disableGutters?: boolean;
-}
-
-export interface ListItemSlots {
-  default: ListItemSlotProps;
-  leadingAdornment: ListItemSlotProps;
-  trailingAdornment: ListItemSlotProps;
-  nestedContent: ListItemSlotProps;
 }
 
 export type ListItemAdornmentProps<T extends EventTarget> = BoxProps<T>;