+++ website/src/routes/list-item/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import { Typography, List, ListItem } from '@uui';
+  import { Typography, List, ListItem, Accordion, Box } from '@uui';
   import DocPage from '../../lib/doc-page.svelte';
   import VariantExample from '../../lib/VariantExample.svelte';
   import InlineCode from '../../lib/inline-code.svelte';
@@ -9,11 +9,19 @@
   import CombinedList from './combined-list.svelte';
   import NestedList from './nested-list.svelte';
   import EventControlList from './event-control-list.svelte';
-  import ListSubheader from './list-subheader.svelte';
-  import ListGroupExample from './list-group-example.svelte';
-  import DeveloperExpierience from './developer-expierience.svelte';
+  import DeveloperExperience from './developer-experience.svelte';
   import LeadingIconList from './leading-icon-list.svelte';
   import TrailingIconList from './trailing-icon-list.svelte';
+  import LIST_ITEM_PROPS from '@uui/List/ListItem.types.d.ts?raw';
+  import AVATAR_EXAMPLE from './avatar-list.svelte?raw';
+  import CHECKBOX_EXAMPLE from './checkbox-list.svelte?raw';
+  import COMBINED_EXAMPLE from './combined-list.svelte?raw';
+  import EVENT_CONTROL_EXAMPLE from './event-control-list.svelte?raw';
+  import LEADING_ICON_EXAMPLE from './leading-icon-list.svelte?raw';
+  import NESTED_EXAMPLED from './nested-list.svelte?raw';
+  import SIMPLE_EXAMPLE from './simple-list.svelte?raw';
+  import TRAILING_ICON_EXAMPLE from './trailing-icon-list.svelte?raw';
+  import AccordionCodeWrapper from '../../lib/accordion-code-wrapper.svelte';
 </script>
 
 <DocPage>
@@ -29,19 +37,7 @@
     <InlineCode>ListItem</InlineCode> has several optional props along with one required (text).
   </Typography>
 
-  <VariantExample
-    code={`
-
-export interface ListItemProps<T extends EventTarget> extends BoxProps<T> {
-  text: string;
-  secondaryText?: string;
-  selected?: boolean;
-  divider?: boolean;
-  disabled?: boolean;
-  disableGutters?: boolean;
-}
-`}
-  />
+  <VariantExample code={LIST_ITEM_PROPS} />
 
   <Typography variant="h5">
     Simple:
@@ -51,6 +47,12 @@ export interface ListItemProps<T extends EventTarget> extends BoxProps<T> {
   </Typography>
 
   <SimpleList />
+  <Accordion>
+    <Typography slot="headerContent" variant="h6">Simple List Code</Typography>
+    <Box slot="content">
+      <VariantExample code={SIMPLE_EXAMPLE} />
+    </Box>
+  </Accordion>
 
   <Typography variant="h5">
     Divider
@@ -99,9 +101,9 @@ export interface ListItemProps<T extends EventTarget> extends BoxProps<T> {
   <Typography variant="body1">
     <InlineCode>ListItem</InlineCode> has three named slots:
     <ul>
-      <li>leadingAdornment</li>
-      <li>trailingAdornment</li>
-      <li>nestedContent</li>
+      <li>leading-adornment</li>
+      <li>trailing-adornment</li>
+      <li>nested-content</li>
     </ul>
   </Typography>
   <Typography variant="body1">
@@ -124,13 +126,13 @@ export interface ListItemProps<T extends EventTarget> extends BoxProps<T> {
   <VariantExample
     code={` 
   <li ...>
-  <slot name="leadingAdornment" />
+  <slot name="leading-adornment" />
     ....
-  <slot name="trailingAdornment" />
+  <slot name="trailing-adornment" />
 </li>
-{#if selected && $$slots.nested}
+{#if selected && $$slots["nested-content"]}
     <div ...>
-      <slot name="nested" />
+      <slot name="nested-content" />
     </div>
 {/if}
   
@@ -141,27 +143,27 @@ export interface ListItemProps<T extends EventTarget> extends BoxProps<T> {
     With Leading Icon:
     <Typography variant="h6" />
   </Typography>
-
   <LeadingIconList />
+  <AccordionCodeWrapper rawCodeExample={LEADING_ICON_EXAMPLE} header="Leading Icon Code" />
 
   <Typography variant="h5">
     With Trailing Icon:
     <Typography variant="h6" />
   </Typography>
-
   <TrailingIconList />
+  <AccordionCodeWrapper rawCodeExample={TRAILING_ICON_EXAMPLE} header="Trailing Icon Code" />
 
   <Typography variant="h5">With Avatar:</Typography>
-
   <AvatarList />
+  <AccordionCodeWrapper rawCodeExample={AVATAR_EXAMPLE} header="Avatar Code" />
 
   <Typography variant="h5">With Checkbox:</Typography>
-
   <CheckboxList />
+  <AccordionCodeWrapper rawCodeExample={CHECKBOX_EXAMPLE} header="Checkbox Code" />
 
   <Typography variant="h5">Combined:</Typography>
-
   <CombinedList />
+  <AccordionCodeWrapper rawCodeExample={COMBINED_EXAMPLE} header="Combined Code" />
 
   <Typography variant="h5">
     With Nested:
@@ -170,8 +172,8 @@ export interface ListItemProps<T extends EventTarget> extends BoxProps<T> {
       <InlineCode>ListItem</InlineCode> or an entire list.
     </Typography>
   </Typography>
-
   <NestedList />
+  <AccordionCodeWrapper rawCodeExample={NESTED_EXAMPLED} header="Nested Code" />
 
   <Typography variant="h3">Controlling Interaction</Typography>
   <Typography variant="body1">
@@ -184,100 +186,10 @@ export interface ListItemProps<T extends EventTarget> extends BoxProps<T> {
     default. Make sure to <InlineCode>stopPropagation()</InlineCode> if you don't want the parent event
     to fire.
   </Typography>
-
   <EventControlList />
+  <AccordionCodeWrapper rawCodeExample={EVENT_CONTROL_EXAMPLE} header="Event Control Code" />
 
-  <Typography variant="h2">List</Typography>
-  <Typography variant="body1">
-    The <InlineCode>List</InlineCode> component acts as the base wrapper for
-    <InlineCode>ListItem</InlineCode>
-    and is used in several other components such as menus, select drop downs, and navigation drawers.
-  </Typography>
-
-  <Typography variant="h3">Types</Typography>
-  <Typography variant="body1">
-    <InlineCode>List</InlineCode> can receive all HTMLAttributes of HTMLUListElement and has one custom
-    prop, disablePadding, which removes the default horizontal and vertical padding.
-  </Typography>
-
-  <VariantExample
-    code={`
-export interface ListProps extends svelte.JSX.HTMLAttributes<HTMLUListElement> {
-  disablePadding?: boolean;
-}
-    `}
-  />
-
-  <div class="inline-examples">
-    <div>
-      <Typography variant="h5">No Padding</Typography>
-      <List class="demo-list" disablePadding>
-        <ListItem text="List Item" />
-      </List>
-    </div>
-    <div>
-      <Typography variant="h5">Padding</Typography>
-      <List class="demo-list">
-        <ListItem text="List Item" selected />
-      </List>
-    </div>
-  </div>
-
-  <Typography variant="h3">Styles</Typography>
-  <Typography variant="body1">
-    Other than the default padding, <InlineCode>List</InlineCode> comes as a blank canvas. You can control
-    the style by passing it a custom class.
-  </Typography>
-  <List class="demo-list" />
-
-  <VariantExample
-    code={`
-    <List class="demo-list" />
-
-.demo-list {
-    width: 300px;
-    border: 1px solid gray;
-    background-color: var(--surface);
-}
-    `}
-  />
-  <Typography variant="h3">Slots</Typography>
-
-  <Typography variant="body1">
-    <InlineCode>List</InlineCode> utilizes an unnamed slot in which any number of list items are rendered
-    as "children".
-  </Typography>
-
-  <VariantExample
-    code={`
-<ul ...> 
-  <slot />
-</ul>
-`}
-  />
-
-  <Typography variant="h4">Subheader</Typography>
-  <Typography variant="body1">
-    The <InlineCode>ListSubhHeader</InlineCode> can be used with single lists, or as we'll see below
-    in list groups.
-  </Typography>
-
-  <ListSubheader />
-
-  <Typography variant="h2">List Group</Typography>
-  <Typography variant="body1">
-    If you want to wrap a group of lists together under sub headings, you can use the
-    <InlineCode>ListGroup</InlineCode> and <InlineCode>ListSubhHeader</InlineCode> components.
-  </Typography>
-
-  <Typography variant="body1">
-    Note: we pass the demo-list class to the <InlineCode>ListGroup</InlineCode> so that the containing
-    foundation is styled.
-  </Typography>
-
-  <ListGroupExample />
-
-  <DeveloperExpierience />
+  <DeveloperExperience />
 </DocPage>
 
 <style lang="scss" global>