+++ website/src/routes/list-item/event-control-list.svelte
@@ -1,14 +1,5 @@
 <script lang="ts">
-  import VariantExample from '../../lib/VariantExample.svelte';
-  import {
-    Accordion,
-    Box,
-    Typography,
-    List,
-    ListItem,
-    ListItemAdornment,
-    ListItemCheckbox
-  } from '@uui';
+  import { List, ListItem, ListItemAdornment, ListItemCheckbox } from '@uui';
 
   let selected = false;
   let exampleOnclick = (index?: number) => {
@@ -26,7 +17,7 @@
     on:click={() => exampleOnclick(0)}
     {selected}
   >
-    <ListItemAdornment slot="trailingAdornment">
+    <ListItemAdornment slot="trailing-adornment">
       <ListItemCheckbox checked={selected} />
     </ListItemAdornment>
   </ListItem>
@@ -36,7 +27,7 @@
     on:click={() => exampleOnclick()}
   >
     <ListItemAdornment
-      slot="trailingAdornment"
+      slot="trailing-adornment"
       on:click={(e) => {
         e.stopPropagation();
         alert('Propagation stopped. No parent event fired.');
@@ -46,40 +37,3 @@
     </ListItemAdornment>
   </ListItem>
 </List>
-
-<Accordion>
-  <Typography slot="headerContent" variant="h6">Event Control Code</Typography>
-  <Box slot="content">
-    <VariantExample
-      code={`
-  <List class="demo-list">
-  <ListItem
-    text="Click Me 1"
-    secondaryText="Checkbox with same click."
-    on:click={() => exampleOnclick(0)}
-    {selected}
-  >
-    <ListItemAdornment slot="trailingAdornment">
-      <ListItemCheckbox checked={selected} />
-    </ListItemAdornment>
-  </ListItem>
-  <ListItem
-    text="Click Me 2"
-    secondaryText="Checkbox with separate click."
-    on:click={() => exampleOnclick()}
-  >
-    <ListItemAdornment
-      slot="trailingAdornment"
-      on:click={(e) => {
-        e.stopPropagation();
-        alert('Propagation stopped. No parent event fired.');
-      }}
-    >
-      <ListItemCheckbox />
-    </ListItemAdornment>
-  </ListItem>
-</List>
-    `}
-    />
-  </Box>
-</Accordion>