+++ website/src/routes/list-item/leading-icon-list.svelte
@@ -1,51 +1,22 @@
 <script>
-  import VariantExample from '$lib/VariantExample.svelte';
-
-  import { List, ListItem, ListItemAdornment, Accordion, Box, Typography } from '@uui';
+  import { List, ListItem, ListItemAdornment } from '@uui';
 </script>
 
 <List class="demo-list">
   <ListItem text="Sent Mail">
-    <ListItemAdornment slot="leadingAdornment" class="material-symbols-outlined">
+    <ListItemAdornment slot="leading-adornment" class="material-symbols-outlined">
       send
     </ListItemAdornment>
   </ListItem>
   <ListItem text="Drafts">
-    <ListItemAdornment slot="leadingAdornment" class="material-symbols-outlined">
+    <ListItemAdornment slot="leading-adornment" class="material-symbols-outlined">
       drafts
     </ListItemAdornment>
   </ListItem>
 
   <ListItem text="Inbox" selected>
-    <ListItemAdornment slot="leadingAdornment" class="material-symbols-outlined">
+    <ListItemAdornment slot="leading-adornment" class="material-symbols-outlined">
       inbox
     </ListItemAdornment>
   </ListItem>
 </List>
-
-<Accordion>
-  <Typography variant="h6" slot="headerContent">Leading Icon Code</Typography>
-  <Box slot="content">
-    <VariantExample
-      code={`
-<List class="demo-list">
-  <ListItem text="Sent Mail">
-    <ListItemAdornment slot="leadingAdornment" class="material-symbols-outlined">
-      send
-    </ListItemAdornment>
-  </ListItem>
-  <ListItem text="Drafts">
-    <ListItemAdornment slot="leadingAdornment" class="material-symbols-outlined">
-      drafts
-    </ListItemAdornment>
-  </ListItem>
-  <ListItem text="Inbox" selected>
-    <ListItemAdornment slot="leadingAdornment" class="material-symbols-outlined">
-      inbox
-    </ListItemAdornment>
-  </ListItem>
-</List>
-`}
-    />
-  </Box>
-</Accordion>